Refactor: response tmpl
