+++ config.yaml
@@ -1,5 +1,5 @@
 telegram:
-  token: 6570837276:AAENP4XX-GB3EkDLiu51WRZQVObkrALkpcQ  # test_bot
+  token: xx  # test_bot
 
 ticker:
   duration: 30s
