+++ notifier/response.tmpl
@@ -1,0 +1,9 @@
+*BUS INFORMATION*
+
+*Coach No:*     _{{ .CoachNo }}_
+*Route Name:*     _{{ .RouteName }}_
+*Available Seats:*     _{{ .AvailableSeats }}_
+*Start Counter:*     _{{ .StartCounter }}_
+*Departure Time:*     _{{ .DepartureTime }}_
+*Company Name:*     _{{ .CompanyName }}_
+*Fare:*     _{{ .Fare }}_
