+++ notifier/tgbot_cmd.go
@@ -42,7 +42,7 @@ func (bot *TelegramBot) CheckCMD(messageText string, chatID int64) {
 		return
 	}
 
-	err = ProcessData(resPld, chatID)
+	err = ProcessData(resPld, chatID, source, destination, dateStr)
 	if err != nil {
 		bot.SendMessage(chatID, "Cannot Fetch Data")
 		return
